fix(recurring): wire providers + router, enforce Decimal for currency
- Add recurringCandidatesProvider + recurringRulesProvider to setall_providers.dart
- Add /recurring route to app_router.dart
- Change RecurringCandidate.amount from double to Decimal
- Detection service emits Decimal from real expense data (closest-to-modal)
- Screen displays via formatAmountForCurrency, persists amount as numeric string
- Confirmed rule display uses Decimal.tryParse for parsing consistency

+++ lib/features/recurring/data/recurring_candidate.dart
@@ -1,3 +1,5 @@
+import 'package:decimal/decimal.dart';
+
 /// A detected recurring-payment candidate before the user confirms/dismisses.
 class RecurringCandidate {
   const RecurringCandidate({
@@ -12,7 +14,7 @@ class RecurringCandidate {
   });
 
   final String description;
-  final double amount;
+  final Decimal amount;
   final String currency;
   final String category;
   final int intervalDays;
